+++ src/main/java/dev/redstudio/witherconfig/config/WitherConfigConfig.java
@@ -25,6 +25,7 @@ public class WitherConfigConfig {
         public double movementSpeed = 0.6; // The "default movement speed of the Wither, it's following speed is "added" to it"
         public double followRange = 40; // Range for the Wither to look for it's target to follow it
         public double armor = 4;
+        public double verticalMovementSpeedMultiplier = 1;
 
         public float unarmoredFlyHeight = 5;
         public float followDistance = 9;