+++ Changelog.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - Boat Jail Fix
   - Boat Jail Fix Tick Delay
   - Boat Jail Fix Range
+  - Vertical Movement Speed Multiplier
 
 ### Fixed